# Serializer perf improvements