… src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoNullable.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandleArray.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandleDictionary.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandleObject.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandlePropertyName.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
… src/System.Text.Json/src/System/Text/Json/Serialization/ReadStackFrame.cs
… src/System.Text.Json/src/System/Text/Json/ThrowHelper.Serialization.cs
… src/System.Text.Json/tests/Serialization/Value.ReadTests.cs