Misc changes to gain ~7% on deserialize and ~3% on serialize for a simple, flat object (as in other recent PRs, the `MicroBenchmarks.Serializers.Json_FromString<Location>` and `MicroBenchmarks.Serializers.Json_ToString<Location>` were used.

There is also a gain for deserializing collections (~6%). There is no gain for serializing collections.

There is also a larger gain for deserializing small payloads which are under 85 bytes (255 / 3) since allocations are stack-based instead of pool-based.

Changes include:
- Avoid the byte[] alloc when deserializing strings (the buffer for transcoding string to byte[]). Now a pool or stack alloc is made.
- Removing several parameters due to exception helper changes. Added support to set an internal `AppendPathInformation` property to true which will cause the try\catch logic to add the path automatically. This avoids having to pass the reader and\or "stack state" to lower-level methods which may need to throw an exception that needs path information as part of its message.
- Remove a temporary struct assignment `WriteStackFrame current = state.Current` that caused a copy-by-value.
- Avoid verifying internal converters in release build for correctness (reading too much or too little).
- Improve the "IsProcessing" logic by using bit flags which helps when comparing more than one `ClassType`.
- Improve `SkipProperty()` by avoiding the direct comparison to the "missing property" singleton.
- Misc other one-line changes

**Deserialize Before**

|         Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|       JSON.NET | 2.554 us | 0.0091 us | 0.0081 us | 2.553 us | 2.540 us | 2.567 us |      0.4771 |           - |           - |              3056 B |
| SystemTextJson | 1.465 us | 0.0088 us | 0.0078 us | 1.464 us | 1.450 us | 1.477 us |      0.1053 |           - |           - |               680 B |

**Deserialize After**

|         Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|       JSON.NET | 2.565 us | 0.0151 us | 0.0134 us | 2.564 us | 2.541 us | 2.590 us |      0.4824 |           - |           - |              3056 B |
| SystemTextJson | 1.358 us | 0.0097 us | 0.0081 us | 1.357 us | 1.341 us | 1.375 us |      0.0701 |           - |           - |               448 B |

**Serialize Before**

|         Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|       JSON.NET | 1.438 us | 0.0106 us | 0.0099 us | 1.438 us | 1.414 us | 1.450 us |      0.2751 |           - |           - |              1736 B |
| SystemTextJson | 1.011 us | 0.0084 us | 0.0079 us | 1.007 us | 1.000 us | 1.024 us |      0.0891 |           - |           - |               584 B |

**Serialize After**

|         Method |       Mean |    Error |   StdDev |     Median |        Min |        Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------- |-----------:|---------:|---------:|-----------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
|       JSON.NET | 1,424.7 ns | 7.607 ns | 7.115 ns | 1,423.7 ns | 1,412.6 ns | 1,439.2 ns |      0.2746 |           - |           - |              1736 B |
| SystemTextJson |   970.9 ns | 7.285 ns | 6.458 ns |   968.5 ns |   964.4 ns |   982.3 ns |      0.0928 |           - |           - |               584 B